+++ oxarchive/__init__.py
@@ -119,7 +119,7 @@ except ImportError:
     OxArchiveWs = None  # type: ignore
     WsOptions = None  # type: ignore
 
-__version__ = "1.9.0"
+__version__ = "1.9.1"
 
 __all__ = [
     # Client